requirements.txt
requests>=2.31.0
kafka-python>=2.0.2
minio>=7.2.0
pyspark>=3.3.2
boto3>=1.26.137








